CAY-961 Change PK generator API to take PK DbAttribute parameter; generate pk based on DbAttribute type
git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/trunk@612246 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/cayenne-jdk1.5-unpublished/src/test/java/org/apache/cayenne/merge/MergerFactoryTest.java
@@ -229,7 +229,7 @@ public class MergerFactoryTest extends CayenneCase {
         column2.setMandatory(false);
         dbEntity.addAttribute(column2);
 
-        DbAttribute column3 = new DbAttribute("ARTIST_ID", Types.INTEGER, dbEntity);
+        DbAttribute column3 = new DbAttribute("ARTIST_ID", Types.BIGINT, dbEntity);
         column3.setMandatory(false);
         dbEntity.addAttribute(column3);
 
@@ -295,7 +295,7 @@ public class MergerFactoryTest extends CayenneCase {
         column2.setMandatory(false);
         dbEntity.addAttribute(column2);
 
-        DbAttribute column3 = new DbAttribute("ARTIST_ID", Types.INTEGER, dbEntity);
+        DbAttribute column3 = new DbAttribute("ARTIST_ID", Types.BIGINT, dbEntity);
         column3.setMandatory(false);
         dbEntity.addAttribute(column3);
